Redefine problem as a set of node pairs
Note that the "old" Single algorithms haven't been updated for this
change and thus cannot be included in configurations anymore. Further,
this fixes a bug in Single_Heur where the selected node pair was not
deterministic over several runs due to the re-use of an
unordered_map.

This makes the "Most" algorithm much more efficient due to a "good"
ordering of the node pairs.