Finder: re-implement find_near, add test, fix bugs
This fixes bugs with nodes being marked as forbidden from previous
calls. This adds a completely new implementation of find_near and
adds variants with callbacks to both the global find() and find_near()
methods. This makes it much easier to test and use them, as no feeder
is required anymore. The test then checks that the global find() and
the local find_near() methods list the same number of forbidden
subgraphs for every node pair for the given graph.