Unify "Most"-implementations, prefer subgraphs with few node pairs
This unifies the Single_Most and Most/Most_Pruned implementations. The
fallback option in Single_Most is now identical to
Most_Pruned. Further, all algorithms always prefer a subgraph with
just a single editable node pair and a subgraph with just two editable
node pairs if one of the two is a node pair with the most number of
forbidden subgraphs.