Single_Most: simplify condition for single
In particular, this simplifies the proof for the running time by
simply allowing length (where length is 4 for P4/C4, 3 for P3 etc.)
editing attempts that are not pruned and then falling back to the
Most_Pruned algorithm. The count is reset after an edit so in the next
recursive call, length calls are available again.