Track and optimize total running time instead of individual ks
The soft time limit is now for the total running time. The hard time
limit is still for each k and can be used to avoid too long
runs. Editors have a separate initialize method now to calculate the
initial lower bound, which is mandatory now. The initial lower bound
is also calculated just once. The initialization time is tracked now
and reported with each k, as well as the total time that includes
initialization and all ks up to that point.