LB_Gurobi: allow compiling with MT editor
This adds a copy constructor that allows to compile the MT editor. The
design of the Gurobi bound still makes it impossible to be executed
with our parallel work stealing. It is possible to change this, but
this will affect performance.

The purpose of this change is to allow compiling the Gurobi bound
together with other bounds that can be used with the MT editor.